backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
fpdf2>=2.8.0
Pillow>=11.0.0
python-multipart>=0.0.12
pydantic>=2.9.0
